pyproject.toml
[tool.poetry]
name = "RespostBot"
version = "0.1.0"
description = ""
authors = ["Pimani <pimanipb@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
discord = "^1.0.1"
numpy = "^1.19.2"
pillow = "^7.2.0"
image_match= "^1.1.2"
scipy = "^1.5.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
